RenderTest : Add new base class to consolidate testing of Render nodes
This contains some initial tests scraped together from the renderer-specific tests that existed already. Those tests will now run for all renderers.
